api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
black==23.11.0
cffi==1.16.0
click==8.1.7
cryptography==3.4.8
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==1.7.1
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
websockets==12.0
